configure dashboard

+++ routes/web.php
@@ -15,41 +15,43 @@ use App\Http\Controllers\UsersController;
 |
 */
 
-Route::get('/', function () {
-    return Inertia::render('Welcome', [
-        'canLogin' => Route::has('login'),
-        'canRegister' => Route::has('register'),
-        'laravelVersion' => Application::VERSION,
-        'phpVersion' => PHP_VERSION,
-    ]);
+// Route::get('/', function () {
+//     return Inertia::render('Welcome', [
+//         'canLogin' => Route::has('login'),
+//         'canRegister' => Route::has('register'),
+//         'laravelVersion' => Application::VERSION,
+//         'phpVersion' => PHP_VERSION,
+//     ]);
+// });
+
+// Route::get('/about', function () {
+//     return Inertia::render('About', [
+//         'canLogin' => Route::has('login'),
+//         'canRegister' => Route::has('register'),
+//         'laravelVersion' => Application::VERSION,
+//         'phpVersion' => PHP_VERSION,
+//     ]);
+// });
+
+// Route::get('/contact', function () {
+//     return Inertia::render('Contact', [
+//         'canLogin' => Route::has('login'),
+//         'canRegister' => Route::has('register'),
+//         'laravelVersion' => Application::VERSION,
+//         'phpVersion' => PHP_VERSION,
+//     ]);
+// });
+
+Route::middleware(['auth:sanctum', 'verified'])->get('/', function() {
+    return redirect('/dashboard');
 });
-
-Route::get('/about', function () {
-    return Inertia::render('About', [
-        'canLogin' => Route::has('login'),
-        'canRegister' => Route::has('register'),
-        'laravelVersion' => Application::VERSION,
-        'phpVersion' => PHP_VERSION,
-    ]);
-});
-
-Route::get('/contact', function () {
-    return Inertia::render('Contact', [
-        'canLogin' => Route::has('login'),
-        'canRegister' => Route::has('register'),
-        'laravelVersion' => Application::VERSION,
-        'phpVersion' => PHP_VERSION,
-    ]);
-});
-
-
 // Users
-Route::get('/users', [UsersController::class, 'index'])->name('users.index');
-Route::get('/users/create', [UsersController::class, 'create'])->name('users.create');
-Route::post('/users', [UsersController::class, 'store'])->name('users.store');
-Route::get('/users/{user}/edit', [UsersController::class, 'edit'])->name('users.edit');
-Route::patch('/users/{user}', [UsersController::class, 'update'])->name('users.update');
-Route::delete('/users/{user}', [UsersController::class, 'destroy'])->name('users.destroy');
+Route::middleware(['auth:sanctum', 'verified'])->get('/users', [UsersController::class, 'index'])->name('users.index');
+Route::middleware(['auth:sanctum', 'verified'])->get('/users/create', [UsersController::class, 'create'])->name('users.create');
+Route::middleware(['auth:sanctum', 'verified'])->post('/users', [UsersController::class, 'store'])->name('users.store');
+Route::middleware(['auth:sanctum', 'verified'])->get('/users/{user}/edit', [UsersController::class, 'edit'])->name('users.edit');
+Route::middleware(['auth:sanctum', 'verified'])->patch('/users/{user}', [UsersController::class, 'update'])->name('users.update');
+Route::middleware(['auth:sanctum', 'verified'])->delete('/users/{user}', [UsersController::class, 'destroy'])->name('users.destroy');
